Some refactoring and change to the reference implementation
